+++ src/components/Global/aos.js
@@ -1,0 +1,9 @@
+import { useEffect } from 'react';
+import Aos from 'aos';
+import 'aos/dist/aos.css';
+
+export const GlobalAOS = () => {
+	useEffect(() => {
+		Aos.init({ duration: 1000 });
+	}, []);
+};
