+++ src/components/Global/themes.js
@@ -1,0 +1,53 @@
+export const darkMode = {
+	background: '#141415',
+	neutral: '#202020e4',
+
+	color: '#f0f0f0',
+	textshadow: '1px 1px 1px black',
+	whitecolor: '#f0f0f0',
+
+	navlogo: '#79c1ff',
+	navlink: '#79c1ff',
+	navborderbottom: 'solid 2px #f0f0f0',
+
+	weight: '500',
+	herounderline: 'underline 3px #79c1ff',
+	borderbottom: 'solid 3px #79c1ff',
+
+	projectlink: '#79c1ff',
+	projectunderline: 'underline #f0f0f0 2px',
+	projectboxshadow: '',
+
+	contactcolor: '#79c1ff',
+	contactunderline: 'underline #f0f0f0 2px',
+
+	footercolor: '#79c1ff',
+	footerunderline: 'underline #f0f0f0 1px',
+};
+
+export const lightMode = {
+	background: 'lightgrey',
+	neutral: '#f0f0f0',
+
+	color: '#333',
+	textshadow: '1px 1px 1px grey',
+	whitecolor: '#333',
+
+	navlogo: '#2BC2C4',
+	navlink: '#2BC2C4',
+	navborderbottom: 'solid 2px #333',
+
+	weight: '',
+	herounderline: 'underline 3px #2BC2C4',
+	borderbottom: 'solid 4px #2BC2C4',
+
+	projectlink: '#2BC2C4',
+	projectunderline: 'underline 2px #444',
+	projectboxshadow: '2px 2px 5px grey',
+
+	contactcolor: '#333',
+	contactunderline: 'underline 2px #2BC2C4',
+
+	footercolor: '',
+	footerunderline: 'underline 2px #2BC2C4',
+};
