+++ src/index.js
@@ -1,7 +1,7 @@
 import React, { Suspense } from 'react';
 import ReactDOM from 'react-dom/client';
-import './index.css';
 import App from './App';
+import './index.css';
 import './i18n.js';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
